Handle the server not sending keepalives within batch limit time.
Adds a timeout that will fire if the keepaive is not seen within
the configured (or default) batch limit time, plus a grace period.

If the TimeoutException is thrown the client will disconnect and
resume the stream by reconnecting. This protects against server
weirdness but also against half-open connections.

The observable's using() setup runs on a single threaded scheduler.
OkHttp needs to close a response on the same thread that opened it
otherwise we'll get errors and may leak resources. Using a single
thread scheduler allows that to happen whereas the default/io/compute
schedulers all use a thead pool that often has the dispose step
trying to close the OkHttp response on the wrong thread.