Place io and processing on different schedulers.
Previously, io and processing were on the same "mono" executor; the
mono exists to ensure the OkHttp response can be closed on the same
thread it was opened on. But using it for processing as well as io
can result in contended pipelines where the processing is starved
as io work dominates.

This creates a new mono that is given to processing. In principle
(and functionally), a regular rx computation scheduler works, but
a single thread for now makes it easier to reason about offset
checkpointing and ordered processing of stream batches.