Remove the option to configure the JSON provider.
Internally, except for one scenario, API handling is abstracted
away from Gson and could use any provider. There's arguably little
value in allowing pluggable JSON support for a client that's
focused on production/ops behaviour - it just provides a vector for
errors. For example, supplying a JSON option that is not configured
to skip nulls won't work with the server API in certain cases. As
another example cursor JSON in headers should be whitespace
compressed. The internal JSON provider abstraction is kept as that's
useful for organizing the code.