Make UndefinedEventMapped and BusinessEventMapped generic.
This allows undefined and business events to be declared with
a generic type instead of a fixed Map<String, Object>. The
previous map option would force client users to reserialise
and deserialise the map to the domain object they're interested
in using.

Getting this to work such that a use can set their custom event
data into the "data" field and have that information lifted into
the top level JSON of the posted event, and have top level JSON
data received in a stream batch mapped down into the same "data"
requires some ugly serialization and deserialization (serdes)
hacks when those types are detected. That's the nature of the HTTP
API we have to work with. Otherwise we'd either have to, drop the
idea of having support for undefined and business, pass around
non-domain types (strings, map, json trees, bytes, etc)
that the client user would have to serdes themselves, or, just
give up on the API and let users represent the entirety of the
two categories as generics passed in and out of the client. We
choose to provide the types instead because a goal of the client
is to represent the full API.

This is a step towards a unified event model. UndefinedEventMapped,
BusinessEventMapped and DataChangeEvent now all have the same
signature for their "data" field.

For #18