Allow a custom SSL socket factory and trust manager to be used.
Some Nakadi installations will use self-signed certificates that
aren't on the system's trust chain. Typically you can install these
certs into the trust store using keytool; this change allows you
supply them via configuration as an alternative.

If the client builder is supplied a certificatePath, the build
will scan that path for .crt and .pem files and load them before
giving a custom socker factory and trust manager to the underlying
OkHttp client. The path must begin with "file:///" or "classpath:"
to indicate whether the certs are loaded from a file directory or
the classpath. If no certificatePath is supplied, the system
defaults are used.

More generally, the client doesn't support a configuration where cert
checking is disabled and turned off.

For #40.