Simplify event sending methods.
This replaces the list and varargs options with a list and
and a single event send methods. The varargs were handy but
provided too many call options and also causing warnings
because the generic event type wasn't reifable (since a
varargs list is converted to a T[] which wipes the type
information and would be illegal if we actually wrote that
out; but since the compiler is doing it we get a warning
instead).

The constraint that the send methods take a wilcard extending
Event has also been lifted, to allow any type to be used.