Allow scopes and custom scopes to be set for event types.
This uses the default scopes where documented, and a best guess
where not documented for event type scopes. The scopes() method
on the EventTypeResource allows a custom scope to be passed in
and can be changed per request invocation. If there's no custom
scope, the default for the request is applied.

For #48.