Clean up http call chains.
This puts all http call chains inside an observable, which allows them
to be constructed once, and avoid bugs where one side of the retry null
conditional is different to the other, or drifts. The retry is added
to the observable if it's not null and not finished.

The not finished check stops unusable retries being configured. Apart
from avoiding no-op retries, it's indicative of a caller bug where
they're recycling retries across requests (retries have state). This
is flagged as a warning log for now; it may be better to throw an
exception; we'll see.