Return a subscription object on create.
Allow a subscription to be returned on create. This is handier
that parsing a Response for the id field. The Response option
is still available, but it's not the "default" create call.

Also provide underlying support for serializing responses from
HTTP requests with bodies, so we can reuse things for other
calls.

For #72.