Reject retry delays that are less than 1s.
When a client handling a partition goes away, the other running clients
will compete to grab it once they wake up and retry. They work by retrying
every N seconds to see if a partition is available and then fall back to
sleep if none are. The aim here is to avoid any need for special level
handling to manage down the number of client connections by the caller.
It should be ok to run the client across your very large microservices
cluster such that even for just a few partitions, they won't hammer the
broker.

This stops the client accepting a value that could cause too many requests;
picking 1s for now, might need to change later.

For #83.