Extend RetryPolicy to support max time backoff.
This provides an option to bound retries by time regardless of the number of
attempts. Event producers that want to bound the total amount of time sending
an event before giving up, and the subscription stream checkpointer are the
main examples in the client that would find it useful to exist based on time
elapse rather than number of attempts.

For #81.