Handle retries with non-throwing requests.
The retry mechanism uses exceptions to drive a retry attempt. This
allows retries to work with non-throwing request() calls, by
internally converting them to throwing requests and tracking the
response. If the retry is exhausted the captured response is
returned instead of throwing the error, which is logged instead.

The exceptions that can cause retries are also given a common
HttpException to make catching the last failed retry easier.

A reasonable question is whether retries should be allowed with
non-throwing requests. For now we'll do that as it's the least
surprising option (probably).