Ignores batch limit values set to 0 or less.
A batch limit of 0 isn't a useful value despite the the API definition
describing it. Instead the server never sends events if batch limit is 0
which is a silent bug. This changes the client to ignore 0 (or lower)
values for batch limit. A future pr prior to 1.0.0 might result in an
exception being thrown but for now ignoring the value keeps with the
current client behaviour.

For #125.