Upgrades stream processors to rxJava2.
Changes the consumers to use RxJava2 internally. There are no interface changes,
but some subtleties and improvements around behaviour:

- RxJava2 has better backpressure support via its Flowable construct. This
is now used in place of Observable and should result in better responsiveness
to the StreamObserver.requestBackPressure request

- Shutdown is graceful but asynchronous. It can be a while before the consumer
stream is suspended after StreamProcessor.stop() is called but for
now, the stop() calls is non-blocking.

- StreamObserver.onCompleted is called properly for a bounded stream request.