Replaces retry handler using FlowableTransformer with a Flowable.
Since rxjava2 StreamConnectionRetry.retryWhenWithBackoffTransformer is
failing to release each event extracted from the buffered reader parsing
the server stream. Instead it buffers lines until the server timesout
the checkpointer and disconnects the client. This causes the consumed
lines to be given to the subscriber which will fail to checkpoint due
to the sesssion id being stale (the session id is recycled on each
connection). The overall retry procedure works, but when its bound to
the flowable it affects the flowable's line string to map to batch
record step stopping a per line emission.

StreamConnectionRetryFlowable is a working replacement that allows lines
to be emitted. It uses the same retry logic as StreamConnectionRetry,
but doesn't process the flowable using zipWith -> flatMap, it just uses
a flatMap.