Allows 422 exceptions from offset commits to be suppressed.
When the server sees a stale sesssion id, it returns 422 which is
converted to an InvalidException and thrown. Session ids become stale
because the server and consumer have disconnected (the server issues a
new session id on each connection and won't honor previous session ids
when submitted with the offset commit). InvalidExceptions are treated
as non-recoverable by the client and cause the overall stream pipeline
to exist stopping the consumer. They're non-recoverable by default because
what to when you can't checkpoint is an application decision.

However many end users have been happy to ignore the exception and continue
working with a resumed stream. This change allows 422s to be suppressed on
a per checkpoint basis which stops the pipeline from exiting (it will instead
recover and resume the stream). This should support less convoluted ways to
trap and handle the exception by end-users.

For #117.