Allows batch offset data to be routed to a subscriber.
The default StreamOffsetObserver behaviour as supplied via
StreamBatchRecord.streamOffsetObserver when called via a StreamObserver
is to block until it returns. This is a sane default as it sequences batch
processing with observers that are performing operations like checkpointing,
where out of order checkpoints would result in incorrect stream positioning.
For example, the SubscriptionOffsetObserver, the default used for subscription
consumers and which calls the API to commit cursors, blocks on the HTTP
response.

It is useful for some cases to allow cursor offsets to be handed off
in sequential order for deferred processing instead of being called directly.
It's also useful to allow more than one observer to see and process cursor
offsets. SubscriptionOffsetPublisher allows both to happen as follows:

- Internally it uses an RxJava Flowable/PublishProcessor combination to emit
each StreamCursorContext received via PublishProcessor.onNext, which is picked
up by the Flowable.

- One or more StreamOffsetObservers can subscribe to StreamCursorContext objects
as a stream via SubscriptionOffsetPublisher.subscribe. This are attached to
the Flowable.

- The current implementation does not trap errors in susbcribers; an error in
any subscriber will break the RxJava flowable chain for all subscribers. This
is intentional for now until multi-subscriber usecases are better understood.

- There are no configurable backpressure options on the RxJava chain yet.
Again this is intentional until use is better understood.

Note that when configuring the StreamProcessor with a
SubscriptionOffsetPublisher, setting an object on StreamProcessor.checkpointer
is a no-op. Clients that want to use the checkpointer need to set it directly
on their subscriber StreamOffsetObservers. The
`testBuilderSubscriptionOffsetPublisher` test shows an example of that.