Allows a specific cert to be loaded from the classpath.
The existing classpath resolver only works when the classpath is
on the filesystem (it scans for files in the supplied folder).
This allows a specific cert ending in pem or crt to be stated,
which should work better with regular classpath resource resolution.

It's preferable to load certs from the keystore. The file and
classpath options are primarily for letting people develop locally
with self-signed certs.

For #126.