Adds a Gson based JsonSupport helper extension.
This replicates the internal functionality of the client for
serializing and deserializing JSON via Gson that has special
handling for Undefined and Business categories. Having an
extension variant that implements the JsonSupport interface
allows support for using Gson annotated objects with event
data. Because the internal Gson support is shaded it can't
be used with Gson directly since items like its annotations
are repackaged.

The client itself needs to be extended to accept a JsonSupport
implementation, but that will be added in a later pr.

For #182.