Reinterrupt and don't log interrupted from the http layer.
This looks for InterruptedIOException coming up from the
rxjava/okio layers and re-interrupts the thread without
logging the trace.

This error can happen normally when the stream processor
is shutting down its executors but the processor is still
reading via its flowable. The log error is a bit of a
red-herring, so skipping it here.

For #186.