Reduce allocations when marshalling events from a stream.
This removes a conversion from a JsonObject to a String, and uses the
JsonObject directly for object marshalling. It also makes a best
effort to clear down an interm list of temp objects created during
marshalling.

It's not clear if client allocations are the cause of OOMs in #256
but this seems like a reasonable overhead to avoid.

For #256.