Avoids StreamProcessor.stop waiting on start completion.
This removes the wait on start latch from the stop call. In earlier
versions, stop was blocked on start to ensure stop didn't pass
through while the processor was still starting and fail to shutdown the
processor. More recently, changes to start and stop logic seem to
avoid the need for this.