Stops the processor after a successful response with a stream timeout.
This checks in the onCancel phase if the response was 200 OK and
whether the processor was configured with a stream_timeout. If so, it's
assumed the server closed a successful response due to the stream timeout
directive sent to it, and the processor will not retry. The 200 Ok check
allows other failure modes (such as authorization) to be retried.

For #291.