Adds a guard clause for calling onStart when processor is stopping.
When the processor's stop is called, the shutdown cycle is
asynchronous. To perform an orderly shutdown, the retry steps
in StreamConnectionRetryFlowable returns a dummy delay instead
of throwing an exception. When that happens the main pipeline
thinks its being told to perform a retry cycle, which results
in a call to doOnSubscribe, which in turn calls the onStart
method of the configured Observer. This doOnSubscribe doesn't
result in the processor starting again and is a noop, but it's
distracting when logged (it looks like the processor is spinning
up) and could cause side effects if the Observer's onStart code
is stateful.

This change checks if the processor is being shutdown, and if it
is, the call to onStart via doOnSubscribe is short circuited. It
also short circuits the stop call if it's detected the processor's
stop was alreacy called.