Allows a consumer to see the checkpoint response result from the server.
This provides a consumer hook to view the CursorCommitResultCollection
result from the server, along with the initiating StreamCursorContext.
If the CursorCommitResultCollection.items() collection is empty
(a 204) from the server it means the server moved the cursor forward.
If the CursorCommitResultCollection.items() collection is not empty
(a 200) it means the server ignored a stale cursor and has returned
its view on the subscription state. A stale cursor is strongly
indicative of a local processing problem.

This change also adds metrics for the two result forms to the
MetricCollector object and calls the collector based on the server
result.